--- client/blocks/checkout-sessions/checkout-sessions-container.js
@@ -1,27 +1,0 @@
-import {
-	CheckoutProvider,
-	PaymentElement,
-} from '@stripe/react-stripe-js/checkout';
-import React, { useMemo } from 'react';
-import CheckoutForm from 'wcstripe/blocks/checkout-sessions/checkout-form';
-import { loadStripe } from 'wcstripe/blocks/load-stripe';
-
-const stripePromise = loadStripe();
-
-export const CheckoutSessionsContainer = ( props ) => {
-	const { api } = props;
-	const promise = useMemo( () => {
-		return api
-			.checkoutSessionsCreateSession()
-			.then( ( r ) => r.client_secret );
-	}, [ api ] );
-
-	return (
-		<CheckoutProvider
-			stripe={ stripePromise }
-			options={ { clientSecret: promise } }
-		>
-			<PaymentElement />
-		</CheckoutProvider>
-	);
-};
