+++ client/blocks/upe/upe-element.js
@@ -13,7 +13,7 @@ import { getBlocksConfiguration } from 'wcstripe/blocks/utils';
 import Icons from 'wcstripe/payment-method-icons';
 import { initializeCheckoutIcons } from 'wcstripe/blocks/upe/checkout-icons';
 import WCStripeAPI from 'wcstripe/api';
-import { CheckoutSessionsContainer } from 'wcstripe/blocks/checkout-sessions/checkout-sessions-container';
+import { CheckoutForm } from 'wcstripe/blocks/checkout-sessions/checkout-form';
 
 // Initialize checkout icons
 const isAdmin = getBlocksConfiguration()?.isAdmin ?? false;
@@ -68,7 +68,7 @@ export const upeElement = ( paymentMethod, api, upeConfig ) => {
 
 	let paymentMethodComponent;
 	if ( true ) { // if checkout sessions is enabled
-		paymentMethodComponent = <CheckoutSessionsContainer api={ api } />;
+		paymentMethodComponent = <CheckoutForm api={ api } />;
 	} else {
 		paymentMethodComponent = getDeferredIntentCreationUPEFields(
 			paymentMethod,
