+++ includes/class-wc-stripe-checkout-sessions-controller.php
@@ -53,7 +53,7 @@ class WC_Stripe_Checkout_Sessions_Controller {
 		}
 
 		$request = [
-			'ui_mode'              => 'custom',
+			'ui_mode'              => 'embedded',
 			'customer'             => $customer->get_id(),
 			'line_items'           => $line_items,
 			'payment_method_types' => $enabled_payment_methods,
